feat: move type into specific folder
