+++ libs/ledger-live-common/src/families/polkadot/react.ts
@@ -10,7 +10,7 @@ import type {
   PolkadotNomination,
   PolkadotSearchFilter,
   PolkadotAccount,
-} from "@ledgerhq/coin-polkadot/types";
+} from "@ledgerhq/coin-polkadot/types/index";
 const SYNC_REFRESH_RATE = 6000; // 6s - block time
 
 export function usePolkadotPreloadData() {
