+++ libs/ledger-live-common/src/families/polkadot/types.ts
@@ -1,2 +1,2 @@
 // Encapsulate for LLD et LLM
-export * from "@ledgerhq/coin-polkadot/types";
+export * from "@ledgerhq/coin-polkadot/types/index";
